implement form object on micropost creation

+++ app/controllers/microposts_controller.rb
@@ -1,34 +1,32 @@
 class MicropostsController < ApplicationController
-  before_action :user_signed_in?, only: [:create, :destroy]
+  before_action :user_signed_in?, only: %i[create destroy]
   before_action :correct_user, only: [:destroy]
 
   def create
-    @micropost = current_user.microposts.build(micropost_params)
-    if @micropost.save
-      flash[:success] = "Micropost created."
+    @micropost = DelayedMicropostForm.new(micropost_params)
+    if @micropost.submit
+      flash[:success] = 'Micropost created.'
       redirect_to root_url
-    else 
+    else
       render 'static_pages/home'
     end
   end
 
   def destroy
-    if @micropost.present?
-      authorize @micropost.destroy!
-    end
+    authorize @micropost.destroy! if @micropost.present?
     redirect_to root_url
-    flash[:success] = "Micropost deleted."
+    flash[:success] = 'Micropost deleted.'
   end
 
   private
 
   def micropost_params
-    params.require(:micropost).permit(:content, :picture)
+    result = params.require(:micropost).permit(:content, :picture, :user, :is_delayed, :pub_date)
+    result.merge('user' => current_user)
   end
 
   def correct_user
     @micropost = current_user.microposts.find(params[:id])
     redirect_to root_url if @micropost.nil?
   end
-
 end
